+++ ta4j-core/src/main/java/org/ta4j/core/indicators/helpers/ClosePriceDifferenceIndicator.java
@@ -24,8 +24,6 @@
 package org.ta4j.core.indicators.helpers;
 
 import org.ta4j.core.BarSeries;
-import org.ta4j.core.indicators.CachedIndicator;
-import org.ta4j.core.num.Num;
 
 /**
  * Calculates the difference between the current and the previous close price.
@@ -33,40 +31,23 @@ import org.ta4j.core.num.Num;
  * <pre>
  * ClosePriceDifference = currentBarClosePrice - previousBarClosePrice
  * </pre>
+ *
+ * @deprecated Use {@link PriceChangeIndicator} with {@link ClosePriceIndicator}
+ *             instead. This class will be removed in a future release.
  */
-public class ClosePriceDifferenceIndicator extends CachedIndicator<Num> {
+@Deprecated
+public class ClosePriceDifferenceIndicator extends PriceChangeIndicator {
 
     /**
      * Constructor.
      *
      * @param series the bar series
+     * @deprecated Use
+     *             {@code new PriceChangeIndicator(new ClosePriceIndicator(series))}
+     *             instead.
      */
+    @Deprecated
     public ClosePriceDifferenceIndicator(BarSeries series) {
-        super(series);
-    }
-
-    /**
-     * Calculates the difference between the close prices of the current bar and the
-     * previous bar.
-     *
-     * @param index the index of the current bar
-     * @return the difference between the close prices
-     */
-    @Override
-    protected Num calculate(int index) {
-        // Get the close price of the previous bar
-        Num previousBarClosePrice = getBarSeries().getBar(Math.max(0, index - 1)).getClosePrice();
-
-        // Get the close price of the current bar
-        Num currentBarClosePrice = getBarSeries().getBar(index).getClosePrice();
-
-        // Calculate the difference between the close prices
-        return currentBarClosePrice.minus(previousBarClosePrice);
-    }
-
-    /** @return {@code 1} */
-    @Override
-    public int getCountOfUnstableBars() {
-        return 1;
+        super(new ClosePriceIndicator(series));
     }
 }