+++ ta4j-core/src/test/java/org/ta4j/core/indicators/helpers/PriceChangePercentageIndicatorTest.java
@@ -23,7 +23,6 @@
  */
 package org.ta4j.core.indicators.helpers;
 
-import static junit.framework.TestCase.assertEquals;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.ta4j.core.TestUtils.assertNumEquals;
 
@@ -67,9 +66,7 @@ public class PriceChangePercentageIndicatorTest extends AbstractIndicatorTest<In
 
     @Test
     public void indicatorShouldReturnNaNForZeroPreviousValue() {
-        BarSeries seriesWithZero = new MockBarSeriesBuilder().withNumFactory(numFactory)
-                .withData(0, 10, 20)
-                .build();
+        BarSeries seriesWithZero = new MockBarSeriesBuilder().withNumFactory(numFactory).withData(0, 10, 20).build();
         PriceChangePercentageIndicator indicator = new PriceChangePercentageIndicator(
                 new ClosePriceIndicator(seriesWithZero));
 
@@ -78,4 +75,3 @@ public class PriceChangePercentageIndicatorTest extends AbstractIndicatorTest<In
         assertNumEquals(numFactory.hundred(), indicator.getValue(2)); // (20 - 10) / 10 * 100 = 100%
     }
 }
-